+++ packages/docs/components/DataBind/stories/basic.js
@@ -1,0 +1,7 @@
+import { registerComponent } from '@studiometa/js-toolkit';
+import { DataBind, DataComputed, DataEffect, DataModel } from '@studiometa/ui';
+
+registerComponent(DataBind);
+registerComponent(DataComputed);
+registerComponent(DataEffect);
+registerComponent(DataModel);