+++ packages/docs/components/DataComputed/stories/basic.js
@@ -1,0 +1,5 @@
+import { registerComponent } from '@studiometa/js-toolkit';
+import { DataComputed, DataModel } from '@studiometa/ui';
+
+registerComponent(DataComputed);
+registerComponent(DataModel);