+++ packages/docs/components/DataComputed/stories/compute-example.js
@@ -1,0 +1,6 @@
+import { registerComponent } from '@studiometa/js-toolkit';
+import { DataBind, DataComputed, DataModel } from '@studiometa/ui';
+
+registerComponent(DataBind);
+registerComponent(DataComputed);
+registerComponent(DataModel);