+++ packages/docs/components/DataEffect/stories/too-long.js
@@ -1,0 +1,5 @@
+import { registerComponent } from '@studiometa/js-toolkit';
+import { DataEffect, DataModel } from '@studiometa/ui';
+
+registerComponent(DataEffect);
+registerComponent(DataModel);