+++ packages/ui/Figure/FigureShopify.ts
@@ -9,9 +9,10 @@ export interface FigureShopifyProps extends BaseProps {
 }
 
 /**
- * Figure class.
+ * FigureShopify class.
  *
  * Manager lazyloading image sources.
+ * @link https://ui.studiometa.dev/-/components/FigureShopify/
  */
 export class FigureShopify<T extends BaseProps = BaseProps> extends AbstractFigureDynamic<
   T & FigureShopifyProps
@@ -33,7 +34,7 @@ export class FigureShopify<T extends BaseProps = BaseProps> extends AbstractFigu
 
   /**
    * Format the source for Shopify CDN API.
-   * @see https://shopify.dev/docs/api/liquid/filters/image_url
+   * @link https://shopify.dev/docs/api/liquid/filters/image_url
    */
   formatSrc(src: string): string {
     const { crop, step } = this.$options;