+++ packages/ui/Transition/Transition.ts
@@ -4,6 +4,7 @@ import { withTransition } from '../decorators/index.js';
 
 /**
  * Transition class.
+ * @link https://ui.studiometa.dev/-/components/Transition/
  */
 export class Transition<T extends BaseProps = BaseProps> extends withTransition<Base>(Base)<T> {
   /**